third_party/WebKit/LayoutTests/fast/forms/select-max-length.html

<p>This test that setting HTMLSelectElement.length is capped to 10,000, but that you can add additional Option elements by calling add.</p>
<div id="console"></div>
<select id="theSelect"></select>
<script src="../../resources/js-test.js"></script>
<script>
    var sel = document.getElementById('theSelect');
    shouldBe('sel.length', '0');

    debug('Trying: - sel.length = 20000;');
    sel.length = 20000;
    shouldBe('sel.length', '0');

    debug('Trying: - sel.add(new Option, 0);');
    sel.add(new Option, 0);
    shouldBe('sel.length', '1');

    debug('Trying: - sel.length = 0;');
    sel.length = 0;
    shouldBe('sel.length', '0');

    debug('Index setter:');
    shouldBe('sel[20000] = new Option(); sel.length', '0');
    shouldBe('sel.options[20000] = new Option(); sel.length', '0');
</script>
