third_party/WebKit/LayoutTests/fast/forms/select-max-length-expected.txt

CONSOLE WARNING: Blocked to expand the option list to 20000 items.  The maximum list length is 10000.
CONSOLE WARNING: Blocked to expand the option list and set an option at index=20000.  The maximum list length is 10000.
CONSOLE WARNING: Blocked to expand the option list and set an option at index=20000.  The maximum list length is 10000.
This test that setting HTMLSelectElement.length is capped to 10,000, but that you can add additional Option elements by calling add.

PASS sel.length is 0
Trying: - sel.length = 20000;
PASS sel.length is 0
Trying: - sel.add(new Option, 0);
PASS sel.length is 1
Trying: - sel.length = 0;
PASS sel.length is 0
Index setter:
PASS sel[20000] = new Option(); sel.length is 0
PASS sel.options[20000] = new Option(); sel.length is 0
PASS successfullyParsed is true

TEST COMPLETE
